Separer DTO i separate filer, legg dem inn i spesifikk dto pakke
Separer DTO i separate filer, legg dem inn i spesifikk dto pakke
