+++ src/main/kotlin/no/nav/familie/ks/barnehagelister/domene/BarnehagelisteService.kt
@@ -1,11 +1,11 @@
 package no.nav.familie.ks.barnehagelister.domene
 
-import no.nav.familie.ks.barnehagelister.kontrakt.BarnehagelisteStatus
-import no.nav.familie.ks.barnehagelister.kontrakt.KindergartenlistResponse
-import no.nav.familie.ks.barnehagelister.kontrakt.ResponseLinks
 import no.nav.familie.ks.barnehagelister.repository.BarnehagelisterRepository
 import no.nav.familie.ks.barnehagelister.rest.ValideringsfeilException
 import no.nav.familie.ks.barnehagelister.rest.ValideringsfeilInfo
+import no.nav.familie.ks.barnehagelister.rest.dto.BarnehagelisteStatus
+import no.nav.familie.ks.barnehagelister.rest.dto.KindergartenlistResponse
+import no.nav.familie.ks.barnehagelister.rest.dto.ResponseLinksResponseDto
 import no.nav.familie.ks.barnehagelister.task.MottattBarnehagelisteTask
 import no.nav.familie.prosessering.internal.TaskService
 import org.slf4j.LoggerFactory
@@ -52,7 +52,7 @@ class BarnehagelisteService(
                         receivedTime = innsendtListe.opprettetTid,
                         finishedTime = innsendtListe.ferdigTid,
                         links =
-                            ResponseLinks(
+                            ResponseLinksResponseDto(
                                 status = "/api/kindergartenlists/status/${skjemaV1.id}",
                             ),
                     ),
@@ -70,7 +70,7 @@ class BarnehagelisteService(
                     receivedTime = barnehageliste.opprettetTid,
                     finishedTime = barnehageliste.ferdigTid,
                     links =
-                        ResponseLinks(
+                        ResponseLinksResponseDto(
                             status = "/api/kindergartenlists/status/${skjemaV1.id}",
                         ),
                 ),
@@ -92,7 +92,7 @@ class BarnehagelisteService(
                     receivedTime = barnehageliste.opprettetTid,
                     finishedTime = barnehageliste.ferdigTid,
                     links =
-                        ResponseLinks(
+                        ResponseLinksResponseDto(
                             status = "/api/kindergartenlists/status/$transaksjonsId",
                         ),
                 ),
