+++ src/main/kotlin/no/nav/familie/ks/barnehagelister/rest/dto/FormV1RequestDto.kt
@@ -1,0 +1,22 @@
+package no.nav.familie.ks.barnehagelister.rest.dto
+
+import io.swagger.v3.oas.annotations.media.Schema
+import jakarta.validation.Valid
+import jakarta.validation.constraints.NotNull
+import java.util.UUID
+
+// TODO sanitize JSON input according to OWASP
+@Schema(description = "Model for submitting kindergarten lists")
+data class FormV1RequestDto(
+    @Schema(
+        description = "Identifier for the submission, generated by the submitter",
+        example = "19375e59-0f07-4c9b-a7bb-6f30fb43819b",
+    )
+    @field:NotNull
+    val id: UUID,
+    @field:Valid
+    val kindergartens: List<KindergartenRequestDto>?,
+    @field:NotNull
+    @field:Valid
+    val listInformationRequestDto: ListInformationRequestDto,
+)
