+++ src/main/kotlin/no/nav/familie/ks/barnehagelister/rest/dto/ListInformationRequestDto.kt
@@ -1,0 +1,19 @@
+package no.nav.familie.ks.barnehagelister.rest.dto
+
+import io.swagger.v3.oas.annotations.media.Schema
+import jakarta.validation.constraints.NotBlank
+import jakarta.validation.constraints.Pattern
+import jakarta.validation.constraints.Size
+import java.time.YearMonth
+
+data class ListInformationRequestDto(
+    @field:NotBlank
+    @field:Size(min = 1, max = 200)
+    val municipalityName: String,
+    @field:NotBlank
+    @field:Size(min = 4, max = 4, message = "Municipality number must have 4 digits")
+    @field:Pattern(regexp = "^[0-9]+(\\.[0-9]+)?$", message = "Municipality number must be a numeric field")
+    val municipalityNumber: String,
+    @Schema(type = "string", format = "yearmonth", example = "2024-09")
+    val submissionForYearMonth: YearMonth,
+)
