+++ src/test/kotlin/no/nav/familie/ks/barnehagelister/rest/UnprotectedBarnehagelisteController.kt
@@ -3,8 +3,8 @@ package no.nav.familie.ks.barnehagelister.rest
 import jakarta.servlet.http.HttpServletRequest
 import no.nav.familie.ks.barnehagelister.domene.BarnehagelisteService
 import no.nav.familie.ks.barnehagelister.domene.mapTilSkjemaV1
-import no.nav.familie.ks.barnehagelister.kontrakt.FormV1
-import no.nav.familie.ks.barnehagelister.kontrakt.KindergartenlistResponse
+import no.nav.familie.ks.barnehagelister.rest.dto.FormV1RequestDto
+import no.nav.familie.ks.barnehagelister.rest.dto.KindergartenlistResponse
 import no.nav.security.token.support.core.api.Unprotected
 import org.springframework.context.annotation.Profile
 import org.springframework.http.ResponseEntity
@@ -19,10 +19,11 @@ class UnprotectedBarnehagelisteController(
 ) : BarnehagelisterController {
     @Unprotected
     override fun mottaBarnehagelister(
-        formV1: FormV1,
+        formV1RequestDto: FormV1RequestDto,
         bindingResult: BindingResult,
         request: HttpServletRequest,
-    ): ResponseEntity<KindergartenlistResponse> = barnehagelisteService.mottaBarnehagelister(formV1.mapTilSkjemaV1(), bindingResult)
+    ): ResponseEntity<KindergartenlistResponse> =
+        barnehagelisteService.mottaBarnehagelister(formV1RequestDto.mapTilSkjemaV1(), bindingResult)
 
     @Unprotected
     override fun status(
